Fix all findings from CROSS_FILE_AUDIT.md
Dead code (12 orphaned functions removed + their now-unused helpers):
- regenerateDeviceId, isCaptivePortalDNSRunning, getRtcTimeString,
  getStorageBarColor, generateDatalogFileOptions, countDatalogFiles,
  formatFileSize, getUsbReservedPins, getProfileById, listProfilesCount,
  FlowRunLogger::isRunning/runStartEpoch (+ _escHtml, <vector> in StorageManager)

Contract mismatches:
- CM-1: CsvLogger::_ensureDir()/begin() now return bool; StorageTask drops to
  drain-only mode instead of silently dropping rows when mkdir fails.
- CM-2: /api/data 'limit' parsed/validated as signed long before size_t cast.
- CM-3: maxRetries() promoted to int once for the retry-loop bound.
- CM-5: untrusted spool 'q' clamped to a defined SensorQuality enumerator.
- CM-4: ExportTask batchCount is size_t (matches sendAll). (SensorManager's
  int n is correct as-is: readAll() returns -1 as an error sentinel.)

Memory lifecycle:
- ML-1: /upload allocates UploadCtx once and registers onDisconnect before
  any validation early-return, so an abort during validation can't leak it.
- ML-2: /import_settings registers its disconnect cleaner immediately after
  allocation, before reserve().
- ML-3: getModeDisplay/getNetworkDisplay kept returning String by value on
  purpose — they feed JsonDocument, which links const char* without copying,
  so const char* would dangle. Documented rather than regressed.

Header bloat:
- Moved <HTTPClient.h> out of 4 exporter headers into their .cpp.
- Moved <Wire.h> out of 7 sensor plugin headers into their .cpp.
- Globals.h: dropped heavy <LittleFS.h>/<SD.h> (kept <FS.h>); added direct
  includes to the two real object users (ESP_Logger.ino, WebServer.cpp).
- Forward-declared fs::FS in StorageManager.h and StorageTask.h.
- Removed redundant <ArduinoJson.h> from WebServer.h (via JsonResponse.h).

+++ src/tasks/ExportTask.cpp
@@ -14,7 +14,7 @@ void exportTaskFunc(void* /*param*/) {
 
     // EXPORT_BATCH_SIZE / EXPORT_FLUSH_INTERVAL_MS are configured in setup.h.
     SensorReading batch[EXPORT_BATCH_SIZE];
-    int           batchCount  = 0;
+    size_t        batchCount  = 0;   // CM-4: match sendAll(…, size_t count)
     uint32_t      lastFlushMs = millis();
 
     SensorReading r;
